refactor(cli): move action services to owning packages
